move trait body gen code to utils
